ai-creator-system/pyproject.toml
[tool.poetry]
name = "ai-creator-system"
version = "0.1.0"
description = "Fully automated AI creator system (Hashnode + YouTube assets + socials)"
authors = ["Irshad <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
pydantic = "^2.9.0"
pydub = "^0.25.1"
pyttsx3 = "^2.90"
Pillow = "^10.4.0"
tenacity = "^9.0.0"
loguru = "^0.7.2"
markdown = "^3.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"